requirements.txt
SpeechRecognition
gTTS
wikipedia
requests
pyyaml
#pyaudio #requires portaudio library (and portaudio-32bit)
#web.py

# Weather
git+https://github.com/jtasker/python-weather-api

# sudo pip install -U setuptools